Move metadata related models to their own package
