+++ plugins/woocommerce/src/main/kotlin/org/wordpress/android/fluxc/model/ProductWithMetaData.kt
@@ -1,5 +1,7 @@
 package org.wordpress.android.fluxc.model
 
+import org.wordpress.android.fluxc.model.metadata.WCMetaData
+
 data class ProductWithMetaData(
     val product: WCProductModel,
     val metaData: List<WCMetaData> = emptyList()
