+++ plugins/woocommerce/src/test/java/org/wordpress/android/fluxc/network/rest/wpcom/wc/product/StripProductMetaDataTest.kt
@@ -5,9 +5,9 @@ import com.google.gson.JsonPrimitive
 import org.assertj.core.api.Assertions
 import org.junit.Test
 import org.wordpress.android.fluxc.model.StripProductMetaData
-import org.wordpress.android.fluxc.model.WCMetaData
-import org.wordpress.android.fluxc.model.WCMetaData.AddOnsMetadataKeys
-import org.wordpress.android.fluxc.model.WCMetaDataValue
+import org.wordpress.android.fluxc.model.metadata.WCMetaData
+import org.wordpress.android.fluxc.model.metadata.WCMetaData.AddOnsMetadataKeys
+import org.wordpress.android.fluxc.model.metadata.WCMetaDataValue
 import org.wordpress.android.fluxc.model.WCProductModel.SubscriptionMetadataKeys
 
 class StripProductMetaDataTest {
